Reorganise ModelStub into stubs package
