+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -22,7 +22,6 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.Tasker;
-import seedu.address.model.consultation.Consultation;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
 import seedu.address.model.person.Person;
 import seedu.address.testutil.EditPersonDescriptorBuilder;
